Move model tests from phases/test_config to config/test_models; delete orphaned file
